move qr inside nonlinear solvers
