--- docs/source/components/qr.rst
@@ -1,6 +1,0 @@
-``qr``
-======
-
-Header ``<pressio/qr.hpp>``
-
-Public namespace: ``pressio::qr``
