+++ include/pressio/solvers_nonlinear/impl/qr/qr_base_classes.hpp
@@ -2,7 +2,7 @@
 //@HEADER
 // ************************************************************************
 //
-// qr_out_of_place_base.hpp
+// qr_in_place_base.hpp
 //                     		  Pressio
 //                             Copyright 2019
 //    National Technology & Engineering Solutions of Sandia, LLC (NTESS)
@@ -46,11 +46,55 @@
 //@HEADER
 */
 
-#ifndef QR_BASE_QR_OUT_OF_PLACE_BASE_HPP_
-#define QR_BASE_QR_OUT_OF_PLACE_BASE_HPP_
+#ifndef QR_BASE_QR_IN_PLACE_BASE_HPP_
+#define QR_BASE_QR_IN_PLACE_BASE_HPP_
 
 namespace pressio{ namespace qr{
 
+template <typename T, typename Q_T, typename Enable = void>
+struct is_legitimate_vector_type_for_qr_project : std::false_type {};
+
+template <typename T, typename Q_t>
+struct is_legitimate_vector_type_for_qr_project<T, Q_t,
+   std::enable_if_t<
+       ::pressio::Traits<T>::rank == 1 and
+     // the vector type should be from same package as Q
+     (false
+#ifdef PRESSIO_ENABLE_TPL_EIGEN
+     or (::pressio::is_native_container_eigen<T>::value and ::pressio::is_native_container_eigen<Q_t>::value)
+#endif
+#ifdef PRESSIO_ENABLE_TPL_KOKKOS
+     or (::pressio::is_native_container_kokkos<T>::value and ::pressio::is_native_container_kokkos<Q_t>::value)
+#endif
+#ifdef PRESSIO_ENABLE_TPL_TRILINOS
+     or (::pressio::is_vector_tpetra<T>::value and ::pressio::is_multi_vector_tpetra<Q_t>::value)
+     or (::pressio::is_vector_tpetra_block<T>::value and ::pressio::is_multi_vector_tpetra_block<Q_t>::value)
+#endif
+     )
+   >
+> : std::true_type{};
+
+
+template<typename DerivedType, typename MatrixType>
+class QRInPlaceBase
+{
+
+  using this_t = QRInPlaceBase<DerivedType, MatrixType>;
+
+  /* workaround for nvcc issue with templates, 
+  see https://devtalk.nvidia.com/default/topic/1037721/nvcc-compilation-error-with-template-parameter-as-a-friend-within-a-namespace/ */
+  template<typename DummyType> struct dummy{using type = DummyType;};
+  friend typename dummy<DerivedType>::type;
+
+public:
+  void computeThin(MatrixType & A){
+    static_cast<DerivedType &>(*this).computeThinImpl(A);
+  }
+
+private:
+  QRInPlaceBase() = default;
+  ~QRInPlaceBase() = default;
+};
 
 template<typename DerivedType, typename MatrixType, typename QType>
 class QROutOfPlaceBase
@@ -75,7 +119,7 @@ public:
   std::enable_if_t<
     ::pressio::Traits<VectorInType>::rank ==1 and
     ::pressio::Traits<VectorOutType>::rank ==1 and
-    meta::is_legitimate_vector_type_for_qr_project<VectorInType, QType>::value
+    is_legitimate_vector_type_for_qr_project<VectorInType, QType>::value
   >
   applyQTranspose(const VectorInType & vecIn, VectorOutType & vecOut) const{
     static_cast<DerivedType const &>(*this).applyQTransposeImpl(vecIn, vecOut);
@@ -92,8 +136,54 @@ public:
 
   QROutOfPlaceBase() = default;
   ~QROutOfPlaceBase() = default;
+};
 
+
+template<typename DerivedType, typename RType>
+class RFactorBase
+{
+
+  using this_t = RFactorBase<DerivedType, RType>;
+
+  /* workaround for nvcc issue with templates, 
+  see https://devtalk.nvidia.com/default/topic/1037721/nvcc-compilation-error-with-template-parameter-as-a-friend-within-a-namespace/ */
+  template<typename DummyType> struct dummy{using type = DummyType;};
+  friend typename dummy<DerivedType>::type;
+
+public:
+  const RType & cRefRFactor() const {
+    return static_cast<DerivedType const&>(*this).cRefRFactorImpl();
+  }
+
+  RFactorBase() = default;
+  ~RFactorBase() = default;
+
+  mutable std::shared_ptr<RType> Rmat_ = nullptr;
+};
+
+
+template<typename DerivedType>
+class QRSolveBase
+{
+  using this_t = QRSolveBase<DerivedType>;
+
+  /* workaround for nvcc issue with templates, 
+  see https://devtalk.nvidia.com/default/topic/1037721/nvcc-compilation-error-with-template-parameter-as-a-friend-within-a-namespace/ */
+  template<typename DummyType> struct dummy{using type = DummyType;};
+  friend typename dummy<DerivedType>::type;
+
+public:
+  template <typename VectorType>
+  std::enable_if_t<
+   ::pressio::Traits<VectorType>::rank ==1 
+  >
+  solve(const VectorType & rhs, VectorType & y)const {
+    static_cast<DerivedType const &>(*this).solveImpl(rhs, y);
+  }
+
+  QRSolveBase() = default;
+  ~QRSolveBase() = default;
 };
 
 }}//end namespace pressio::qr
-#endif  // QR_BASE_QR_OUT_OF_PLACE_BASE_HPP_
+#endif  // QR_BASE_QR_IN_PLACE_BASE_HPP_
