+++ include/pressio/solvers_nonlinear_gaussnewton.hpp
@@ -54,8 +54,9 @@
 #include "./type_traits.hpp"
 #include "./expressions.hpp"
 #include "./ops.hpp"
-#include "./qr.hpp"
 #include "./solvers_linear.hpp"
+
+#include "./solvers_nonlinear/impl/qr.hpp"
 
 #include "solvers_nonlinear_concepts.hpp"
 #include "solvers_nonlinear/solvers_exceptions.hpp"
