+++ tests/functional_small/CMakeLists.txt
@@ -6,7 +6,6 @@ add_subdirectory(utils)
 add_subdirectory(type_traits)
 add_subdirectory(expressions)
 add_subdirectory(ops)
-add_subdirectory(qr)
 add_subdirectory(solvers_linear)
 add_subdirectory(solvers_nonlinear)
 add_subdirectory(ode_advancers)
