+++ tests/functional_small/solvers_nonlinear/qr/CMakeLists.txt
