+++ tests/functional_small/solvers_nonlinear/qr/fixtures.hpp
@@ -3,7 +3,7 @@
 #define QR_UTEST_FIXTURES_HPP_
 
 #include <gtest/gtest.h>
-#include "pressio/qr.hpp"
+#include "pressio/solvers_nonlinear.hpp"
 #include "qr_r9c4_gold.hpp"
 
 #ifdef PRESSIO_ENABLE_TPL_TRILINOS
