+++ tests/functional_small/solvers_nonlinear/qr/qr_r9c4_gold.hpp
